+++ volumewalker/command.h
@@ -3,12 +3,12 @@
 
 HANDLE CreateCommandHandler(HWND hWnd);
 BOOL CloseCommandHandler(HANDLE hCommand);
-BOOL ForwardCommand(HANDLE hCommand,UINT idCmd);
-HMENU MakeVolumeCommandMenu(HANDLE hCommand);
-VOID InitializeVolumeTools();
 
 namespace CommandHandler
 {
-	BOOL Message(MSG *pmsg);
+	BOOL PreTranslateMessage(MSG *pmsg);
 	INT QueryCmdState(UINT CmdId,INT& State);
+	BOOL NotifyClose(HWND);
+	BOOL ForwardCommand(HANDLE hCommand,UINT idCmd);
+	HMENU MakeVolumeCommandMenu(HANDLE /*hCommand*/);
 };